requirements.txt
python-dotenv
Pillow
requests
google-maps-routing
